Update and simplify demo example
Updated the demo example to use features from the latest version of the
`ratatui` crate and simplified the application model significantly. Note
that the main crate is still compatible with `ratatui` version `0.25.0`,
but the demo example uses features from `0.26.0`.

Moved the optional dependencies on `termion` and `crossterm` to the
`dev-dependencies` section of the `Cargo.toml` file, and made them
mutually exclusive in the demo example. Moved the backend specific code
into seperate modules.

To run the demo, use either of the following commands:

```shell
cargo run --example demo --features crossterm
cargo run --example demo --features termion
```

Added a bacon config file to make it easy to check that the demo
compiles with the `crossterm` and `termion` features enabled.